Add command line parsing
The --debug option sets whether the server prints logging messages.

Note that this doesn't work when it comes to launching secondary
processes when one process is already running.  The secondary processes
are supposed to signal the primary and then exit, but they don't when
HANDLES_COMMAND_LINE is set.  I'm leaving this commit here in case that
bug gets fixed.