Get multiple instances to work together
The official ways to do this (HANDLES_COMMAND_LINE, HANDLES_OPEN) appear
to be completely broken for Python.  So I implement my own method that
checks whether or not it is the primary instance.  The only way I could
get instances to talk with each other is by triggering actions.  So
secondary instances trigger actions after processing their command lines
to set the arguments needed when thy primary instance is activated.