Enable page turns from the keyboard
The original plan was to do this with javascript in the WebView.  But
Monocle uses an iframe whose contents keep being refreshed, so we'd have
to continually replace the handlers.  Next, I tried to create actions to
advance and reverse the pages, but you can't hook up bare arrow keys to
accelerators.  So finally I just handled the keypress by hand.