Switch server to libsoup
This allows an event-driven server running inside the GTK main loop.
Thus, we don't have to muck about with threads.  It also means that we
could tie the server and client more closely together.

In the process, we fixed the on_quit event for the individual windows.

One problem is that the server will respond to request from anywhere,
rather than just from localhost, as it had been.  It'd be nice to fix
this.